Resolve test issues.
