+++ src/components/Viewer/Painting/Painting.test.tsx
@@ -4,7 +4,6 @@ import { act, fireEvent, render, screen } from "@testing-library/react";
 import { IIIFExternalWebResource } from "@iiif/presentation-3";
 import ImageViewer from "src/components/Image";
 import { LabeledIIIFExternalWebResource } from "src/types/presentation-3";
-import Map from "src/components/Map";
 import Painting from "src/components/Viewer/Painting/Painting";
 import Placeholder from "src/components/Viewer/Painting/Placeholder";
 import Player from "src/components/Viewer/Player/Player";
@@ -61,9 +60,6 @@ vi.mock("src/components/Image");
 vi.mocked(ImageViewer).mockReturnValue(
   <div data-testid="mock-image-viewer">ImageViewer</div>,
 );
-
-vi.mock("src/components/Map");
-vi.mocked(Map).mockReturnValue(<div data-testid="mock-map">Map</div>);
 
 vi.mock("src/components/Viewer/Painting/Placeholder");
 vi.mocked(Placeholder).mockReturnValue(
@@ -162,431 +158,6 @@ describe("Painting component", () => {
     expect(screen.queryByTestId("placeholder-toggle")).toBeNull();
     expect(screen.queryByTestId("mock-player")).not.toBeInTheDocument();
     expect(screen.getByTestId("mock-image-viewer")).toBeInTheDocument();
-  });
-
-  it("displays the Map viewer when map mode is enabled and active canvases have navPlace", async () => {
-    const mapVault = new Vault();
-    const activeCanvas = "https://example.org/iiif/canvas/navplace/1";
-    const manifest = {
-      "@context": "http://iiif.io/api/presentation/3/context.json",
-      id: "https://example.org/iiif/manifest/navplace",
-      type: "Manifest",
-      label: { en: ["Map Manifest"] },
-      items: [
-        {
-          id: activeCanvas,
-          type: "Canvas",
-          height: 1000,
-          width: 1000,
-          label: { en: ["Canvas with navPlace"] },
-          navPlace: {
-            type: "FeatureCollection",
-            features: [
-              {
-                type: "Feature",
-                properties: { label: { en: ["Canvas point"] } },
-                geometry: {
-                  type: "Point",
-                  coordinates: [-87.6881, 42.045],
-                },
-              },
-            ],
-          },
-          items: [
-            {
-              id: `${activeCanvas}/page`,
-              type: "AnnotationPage",
-              items: [
-                {
-                  id: `${activeCanvas}/annotation`,
-                  type: "Annotation",
-                  motivation: "painting",
-                  body: painting[0],
-                  target: activeCanvas,
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    };
-    await mapVault.loadManifest("", manifest);
-
-    render(
-      <ViewerProvider
-        initialState={{
-          ...defaultState,
-          activeManifest: manifest.id,
-          configOptions: {
-            ...defaultState.configOptions,
-            map: {
-              enabled: true,
-              fitToData: true,
-              navPlaceLevel: "auto",
-            },
-          },
-          vault: mapVault,
-          visibleCanvases: [{ id: activeCanvas, type: "Canvas" }],
-        }}
-      >
-        <Painting {...defaultProps} activeCanvas={activeCanvas} />
-      </ViewerProvider>,
-    );
-
-    expect(screen.getByTestId("mock-map")).toBeInTheDocument();
-    expect(screen.queryByTestId("mock-image-viewer")).not.toBeInTheDocument();
-    expect(vi.mocked(Map)).toHaveBeenCalledWith(
-      expect.objectContaining({
-        fitToData: true,
-        navPlace: expect.objectContaining({
-          type: "FeatureCollection",
-        }),
-      }),
-      expect.anything(),
-    );
-
-    const mapProps = vi.mocked(Map).mock.calls.at(-1)?.[0];
-    const navPlace = mapProps?.navPlace as GeoJSON.FeatureCollection;
-    expect(navPlace.features[0].properties?.iiifResource).toMatchObject({
-      id: activeCanvas,
-      type: "Canvas",
-      parent: {
-        id: manifest.id,
-        type: "Manifest",
-      },
-    });
-  });
-
-  it("displays Manifest-level navPlace when map mode targets the Manifest", async () => {
-    const mapVault = new Vault();
-    const activeCanvas = "https://example.org/iiif/canvas/manifest-navplace/1";
-    const manifest = {
-      "@context": "http://iiif.io/api/presentation/3/context.json",
-      id: "https://example.org/iiif/manifest/manifest-navplace",
-      type: "Manifest",
-      label: { en: ["Manifest with navPlace"] },
-      summary: { en: ["Manifest-level map summary"] },
-      thumbnail: [{ id: "https://example.org/thumb.jpg", type: "Image" }],
-      homepage: [{ id: "https://example.org/work", type: "Text" }],
-      navPlace: {
-        type: "FeatureCollection",
-        features: [
-          {
-            type: "Feature",
-            properties: { label: { en: ["Pend Oreille River"] } },
-            geometry: {
-              type: "Point",
-              coordinates: [-117.61775, 49.00381],
-            },
-          },
-        ],
-      },
-      items: [
-        {
-          id: activeCanvas,
-          type: "Canvas",
-          height: 1000,
-          width: 1000,
-          label: { en: ["Canvas 1"] },
-          items: [
-            {
-              id: `${activeCanvas}/page`,
-              type: "AnnotationPage",
-              items: [
-                {
-                  id: `${activeCanvas}/annotation`,
-                  type: "Annotation",
-                  motivation: "painting",
-                  body: painting[0],
-                  target: activeCanvas,
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    };
-    await mapVault.loadManifest("", manifest);
-
-    render(
-      <ViewerProvider
-        initialState={{
-          ...defaultState,
-          activeManifest: manifest.id,
-          configOptions: {
-            ...defaultState.configOptions,
-            map: {
-              enabled: true,
-              fitToData: true,
-              navPlaceLevel: "Manifest",
-            },
-          },
-          vault: mapVault,
-          visibleCanvases: [{ id: activeCanvas, type: "Canvas" }],
-        }}
-      >
-        <Painting {...defaultProps} activeCanvas={activeCanvas} />
-      </ViewerProvider>,
-    );
-
-    expect(screen.getByTestId("mock-map")).toBeInTheDocument();
-
-    const mapProps = vi.mocked(Map).mock.calls.at(-1)?.[0];
-    const navPlace = mapProps?.navPlace as GeoJSON.FeatureCollection;
-    expect(navPlace.features[0].properties?.iiifResource).toMatchObject({
-      id: manifest.id,
-      type: "Manifest",
-      label: manifest.label,
-      summary: manifest.summary,
-      thumbnail: "https://example.org/thumb.jpg",
-      homepage: "https://example.org/work",
-    });
-  });
-
-  it("gathers and adapts georef annotations across the manifest for overlays", async () => {
-    const overlayVault = new Vault();
-    const canvas1 = "https://example.org/iiif/canvas/georef/1";
-    const canvas2 = "https://example.org/iiif/canvas/georef/2";
-    const service1 = "https://iiif.example.org/image/2/sheet-1";
-    const service2 = "https://iiif.example.org/image/2/sheet-2";
-
-    const georefBody = (coords: [number, number][]) => ({
-      type: "FeatureCollection",
-      features: coords.map((coordinates, index) => ({
-        type: "Feature",
-        properties: { resourceCoords: [index * 10, index * 10] },
-        geometry: { type: "Point", coordinates },
-      })),
-    });
-
-    const buildCanvas = (
-      canvasId: string,
-      serviceId: string,
-      coords: [number, number][],
-    ) => ({
-      id: canvasId,
-      type: "Canvas",
-      height: 1000,
-      width: 1000,
-      items: [
-        {
-          id: `${canvasId}/page`,
-          type: "AnnotationPage",
-          items: [
-            {
-              id: `${canvasId}/annotation`,
-              type: "Annotation",
-              motivation: "painting",
-              body: {
-                id: `${serviceId}/full/600,/0/default.jpg`,
-                type: "Image",
-                format: "image/jpeg",
-                height: 1000,
-                width: 1000,
-                service: [{ id: serviceId, type: "ImageService2" }],
-              },
-              target: canvasId,
-            },
-          ],
-        },
-      ],
-      annotations: [
-        {
-          id: `${canvasId}/georef-page`,
-          type: "AnnotationPage",
-          items: [
-            {
-              id: `${canvasId}/georef`,
-              type: "Annotation",
-              motivation: "georeferencing",
-              target: {
-                type: "SpecificResource",
-                source: { id: canvasId, type: "Canvas", width: 1000, height: 1000 },
-              },
-              body: georefBody(coords),
-            },
-          ],
-        },
-      ],
-    });
-
-    const manifest = {
-      "@context": "http://iiif.io/api/presentation/3/context.json",
-      id: "https://example.org/iiif/manifest/georef",
-      type: "Manifest",
-      label: { en: ["Georeferenced Sheets"] },
-      items: [
-        buildCanvas(canvas1, service1, [
-          [-87.71, 42.06],
-          [-87.68, 42.06],
-          [-87.68, 42.01],
-        ]),
-        buildCanvas(canvas2, service2, [
-          [-87.78, 42.07],
-          [-87.7, 42.07],
-          [-87.7, 42.08],
-        ]),
-      ],
-    };
-    await overlayVault.loadManifest("", manifest);
-
-    render(
-      <ViewerProvider
-        initialState={{
-          ...defaultState,
-          activeManifest: manifest.id,
-          configOptions: {
-            ...defaultState.configOptions,
-            map: {
-              enabled: true,
-              fitToData: true,
-              showImageOverlay: true,
-              overlayScope: "manifest",
-            },
-          },
-          vault: overlayVault,
-          visibleCanvases: [{ id: canvas1, type: "Canvas" }],
-        }}
-      >
-        <Painting {...defaultProps} activeCanvas={canvas1} />
-      </ViewerProvider>,
-    );
-
-    expect(screen.getByTestId("mock-map")).toBeInTheDocument();
-
-    await vi.waitFor(() => {
-      const mapProps = vi.mocked(Map).mock.calls.at(-1)?.[0];
-      const annotations = mapProps?.georefAnnotations ?? [];
-      expect(annotations).toHaveLength(2);
-    });
-
-    const mapProps = vi.mocked(Map).mock.calls.at(-1)?.[0];
-    const annotations = mapProps?.georefAnnotations ?? [];
-
-    // Each Canvas-sourced annotation is adapted to its painting image service.
-    const sources = annotations
-      .map((annotation) => annotation.target?.source)
-      .sort((a, b) => (a?.id ?? "").localeCompare(b?.id ?? ""));
-    expect(sources[0]).toMatchObject({ id: service1, type: "ImageService2" });
-    expect(sources[1]).toMatchObject({ id: service2, type: "ImageService2" });
-
-    // Overlay options are threaded through.
-    expect(mapProps?.showImageOverlay).toBe(true);
-  });
-
-  it("limits overlay discovery to the visible canvas when overlayScope is 'canvas'", async () => {
-    const overlayVault = new Vault();
-    const canvas1 = "https://example.org/iiif/canvas/scope/1";
-    const canvas2 = "https://example.org/iiif/canvas/scope/2";
-    const service1 = "https://iiif.example.org/image/2/scope-1";
-    const service2 = "https://iiif.example.org/image/2/scope-2";
-
-    const buildCanvas = (canvasId: string, serviceId: string) => ({
-      id: canvasId,
-      type: "Canvas",
-      height: 1000,
-      width: 1000,
-      items: [
-        {
-          id: `${canvasId}/page`,
-          type: "AnnotationPage",
-          items: [
-            {
-              id: `${canvasId}/annotation`,
-              type: "Annotation",
-              motivation: "painting",
-              body: {
-                id: `${serviceId}/full/600,/0/default.jpg`,
-                type: "Image",
-                format: "image/jpeg",
-                height: 1000,
-                width: 1000,
-                service: [{ id: serviceId, type: "ImageService2" }],
-              },
-              target: canvasId,
-            },
-          ],
-        },
-      ],
-      annotations: [
-        {
-          id: `${canvasId}/georef-page`,
-          type: "AnnotationPage",
-          items: [
-            {
-              id: `${canvasId}/georef`,
-              type: "Annotation",
-              motivation: "georeferencing",
-              target: {
-                type: "SpecificResource",
-                source: { id: canvasId, type: "Canvas", width: 1000, height: 1000 },
-              },
-              body: {
-                type: "FeatureCollection",
-                features: [
-                  {
-                    type: "Feature",
-                    properties: { resourceCoords: [0, 0] },
-                    geometry: { type: "Point", coordinates: [-87.7, 42.0] },
-                  },
-                  {
-                    type: "Feature",
-                    properties: { resourceCoords: [10, 10] },
-                    geometry: { type: "Point", coordinates: [-87.6, 42.0] },
-                  },
-                  {
-                    type: "Feature",
-                    properties: { resourceCoords: [20, 20] },
-                    geometry: { type: "Point", coordinates: [-87.6, 42.1] },
-                  },
-                ],
-              },
-            },
-          ],
-        },
-      ],
-    });
-
-    const manifest = {
-      "@context": "http://iiif.io/api/presentation/3/context.json",
-      id: "https://example.org/iiif/manifest/scope",
-      type: "Manifest",
-      label: { en: ["Scope Test"] },
-      items: [buildCanvas(canvas1, service1), buildCanvas(canvas2, service2)],
-    };
-    await overlayVault.loadManifest("", manifest);
-
-    render(
-      <ViewerProvider
-        initialState={{
-          ...defaultState,
-          activeManifest: manifest.id,
-          configOptions: {
-            ...defaultState.configOptions,
-            map: {
-              enabled: true,
-              fitToData: true,
-              showImageOverlay: true,
-              overlayScope: "canvas",
-            },
-          },
-          vault: overlayVault,
-          visibleCanvases: [{ id: canvas1, type: "Canvas" }],
-        }}
-      >
-        <Painting {...defaultProps} activeCanvas={canvas1} />
-      </ViewerProvider>,
-    );
-
-    await vi.waitFor(() => {
-      const mapProps = vi.mocked(Map).mock.calls.at(-1)?.[0];
-      expect(mapProps?.georefAnnotations ?? []).toHaveLength(1);
-    });
-
-    const mapProps = vi.mocked(Map).mock.calls.at(-1)?.[0];
-    expect(mapProps?.georefAnnotations?.[0]?.target?.source).toMatchObject({
-      id: service1,
-    });
   });
 
   it.skip("renders the Choice select dropdown", () => {
